interpreter: Fix variable mutation in block
When mutating a variable from a block, the interpreter would call the
`replace_variable` method, which would get a reference to the variable,
remove it, and then add a new one. However, this method wasn't scope-map
aware: While we were removing the correct variable from the outtermost
scope, we were adding the new one in the inner scope. An easy fix is to
instead mutate the variable we were originally getting.